Update README.md
Corrected the instructions  at line 15 specifying how to log in. I downloaded the entire code pckg on July 25, 2016. But I could not get it the login to work. But today I discovered that User.sql file apparently (?) does not have the change of username which is noted on the commit for that file which was committed 2 months earlier. Now I can login.